pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.3.0",
    "flask-sqlalchemy>=3.1.1",
    "flask>=3.1.2",
    "google-genai>=1.38.0",
    "gunicorn>=23.0.0",
    "numpy>=2.3.3",
    "pandas>=2.3.2",
    "pillow>=11.3.0",
    "pydantic>=2.11.9",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "scikit-learn>=1.7.2",
    "sift-stack-py>=0.9.1",
    "werkzeug>=3.1.3",
    "psycopg2-binary>=2.9.10",
]
